fixed small review requests
